feat(pattern): extract matcher into moq-pattern and @moq/pattern
