+++ js/net/package.json
@@ -17,6 +17,7 @@
 		"release": "bun ../common/release.ts"
 	},
 	"dependencies": {
+		"@moq/pattern": "workspace:^",
 		"@moq/qmux": "^0.3.3",
 		"@moq/signals": "workspace:*",
 		"async-mutex": "^0.5.0",
