+++ js/pattern/tsconfig.json
@@ -1,0 +1,9 @@
+{
+	"extends": "../tsconfig.json",
+	"compilerOptions": {
+		"outDir": "dist",
+		"rootDir": "./src",
+		"types": ["bun"]
+	},
+	"include": ["src"]
+}
