+++ rs/moq-pattern/tests/pattern.json
