+++ rs/moq-token/Cargo.toml
@@ -17,6 +17,7 @@ tokio = ["dep:tokio"]
 aws-lc-rs = { workspace = true }
 base64 = { workspace = true }
 jsonwebtoken = { workspace = true, features = ["aws_lc_rs"] }
+moq-pattern = { workspace = true, features = ["serde"] }
 p256 = "0.14.0"
 p384 = "0.14.0"
 reqwest = { workspace = true, features = ["rustls"], optional = true }
