+++ rs/moq-token/src/lib.rs
@@ -4,6 +4,9 @@
 //! Tokens specify which broadcast paths a client can publish to and consume from.
 //!
 //! See [`Claims`] for the JWT claims structure and [`Key`] for key management.
+//! Path grants use [`Pattern`] from [`moq-pattern`](moq_pattern): the same grammar
+//! `moq-net` re-exports. New minting should construct patterns rather than prefixes;
+//! missing `v` on persisted claims stays v0 prefix semantics.
 
 mod algorithm;
 mod claims;
@@ -20,4 +23,5 @@ pub use claims::*;
 pub use error::*;
 pub use key::*;
 pub use key_id::*;
+pub use moq_pattern::{InvalidPattern, Pattern, Patterns, Segment, Specificity};
 pub use set::*;
